rename back to editor
